Wiping all the PHP 5.3 support away.
