+++ .travis.yml
@@ -5,6 +5,7 @@ php:
   - '5.6'
   - '7.0'
   - '7.1'
+  - '7.2'
   - 'nightly'
 
 script: vendor/bin/phpunit --configuration ./build/travis-ci.xml
