+++ composer.json
@@ -1,22 +1,21 @@
 {
-    "name": "gburtini/distributions",
-    "description": "PHP implementation of a number of statistical probability distributions: normal, beta, gamma, etc.",
-    "authors": [
-        {
-            "name": "Giuseppe Burtini",
-            "email": "joe@truephp.com"
-        }
-    ],
-    "license": "MIT",
-    "autoload": {
-        "psr-0": {
-            "gburtini\\Distributions":        "src"
-        }
-    },
-    "require": {},
-    "require-dev": {
-        "phpunit/phpunit": "4.4.*",
-        "phpunit/php-code-coverage": "^2.2"
+  "name": "gburtini/distributions",
+  "description":
+    "PHP implementation of a number of statistical probability distributions: normal, beta, gamma, etc.",
+  "authors": [
+    {
+      "name": "Giuseppe Burtini",
+      "email": "giuseppe@burtini.ca"
     }
-
+  ],
+  "license": "MIT",
+  "autoload": {
+    "psr-0": {
+      "gburtini\\Distributions": "src"
+    }
+  },
+  "require": {},
+  "require-dev": {
+    "phpunit/phpunit": "^7"
+  }
 }
