--- src/gburtini/Distributions/README.md
@@ -1,6 +1,0 @@
-Confused?
----------
-
-As of 7 December 2015 this directory provides composer-compatibility and pretty autoloading namespaces but *does not* directly provide implementations. Implementations are all in the ``ugly/`` subdirectory in the root of the repository. This is done to provide compatibility and ease-of-use with pre-namespace versions of PHP.
-
-As we have removed official support for PHP < 5.3, this feature may disappear as soon as there is a good reason to.
